MFC r360051: tty: convert tty_lock_assert to tty_assert_locked
A later change, currently being iterated on in D24459, will in-fact change
the lock type to an sx so that TTY drivers can sleep on it if they need to.
Committing this ahead of time to make the review in question a little more
palatable.

tty_lock_assert() is unfortunately still needed for now in two places to
make sure that the tty lock has not been recursed upon, for those scenarios
where it's supplied by the TTY driver and possibly a mutex that is allowed
to recurse.


git-svn-id: https://svn.freebsd.org/base/stable/12@360258 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f