MFC r359642: adduser: allow standard IFS characters in passwords
Notably, the default IFS contains space/tab, thus any leading/trailing
whitespace characters tend to be removed.

Set IFS= for just the read lines to mitigate this, allowing the user to be
less surprised when their leading/trailing spaces weren't actually captured
in the password as they are with other means of setting a user's password.

PR:		245342


git-svn-id: https://svn.freebsd.org/base/stable/12@360390 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f