MFC r360071-r360072: force -fcommon in kernel/modules where needed
This contains some extra bits over the original commits, as things are
slightly different in earlier branches.

r360071:
Allow kernel modules to build with a compiler that defaults to -fno-common

This uses the same approach as r359691.

r360072:
More fixes to build the kernel with a compiler that defaults to -fno-common

Using the same approach as the last commit for the files used by genassym.sh.


git-svn-id: https://svn.freebsd.org/base/stable/12@360496 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f