Make mpentry independent of _start
APs enter the kernel at the same point as the BSP, the _start routine.
They then jump to mpentry, but not before storing the kernel's physical
load address in the s9 register. Extract this calculation into its own
routine, so that APs can be instructed to enter directly from mpentry.

Differential Revision:	https://reviews.freebsd.org/D24495


git-svn-id: https://svn.freebsd.org/base/head@360551 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f