MFC r355940:
Move all sources from the llvm project into contrib/llvm-project.

This uses the new layout of the upstream repository, which was recently
migrated to GitHub, and converted into a "monorepo".  That is, most of
the earlier separate sub-projects with their own branches and tags were
consolidated into one top-level directory, and are now branched and
tagged together.

Updating the vendor area to match this layout is next.


git-svn-id: https://svn.freebsd.org/base/stable/11@360660 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f