MFC r350027, r350028: Let traceroute6 not ignore some ICMPv6 packets
MFC rr350027:
Let packet_op() explicitly return the type and code instead of doing
this implicitly by encoding it in a number space.
No functional change intended.
This is done as a preparation to add support for ICMPv6 mesages
indicating a parameter problem related to the next header.

MFC r350028:
Add support for ICMPv6 messages indicating a parameter problem related
to an unrecognized next header.


git-svn-id: https://svn.freebsd.org/base/stable/11@360709 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f