lualoader: fix loader.lua mis-merge
This mis-merge came from r360599; r331321 introduced core.isUEFIBoot but
will not get merged back to stable/11 due to being somewhat disruptive. I
had resolved the conflict once, but had to re-do the merge locally due to
another conflict and failed to resolve it the second time.

This mistake is explosive to those trying to run lualoader on stable/11, but
is mitigated by the fact that lualoader is not the default here.

PR:		246291


git-svn-id: https://svn.freebsd.org/base/stable/11@360795 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f