src.opts.mk: update BINUTILS options and add comments
BINUTILS is needed only for ports, and will be disabled once the failing
ports are addressed (likely by growing a binutils dependency).

BINUTILS_BOOTSTRAP is needed only on amd64, for skein_block_asm.s. There
is no need to enable it on i386.

This will all be removed before FreeBSD 13.0.


git-svn-id: https://svn.freebsd.org/base/head@360815 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f